♻️ refactor: cursor 계산 함수 util로 분리
